Fix issue with some content-types like text/plain
Python's string.lstrip function is not doing what
is expected at the specific position, resulting
in the fact that 'content-type: text/plain'
is modified to 'xt/plain' instead of 'text/plain'.
The test is modified to check for different
content-types as 'image/png' is by chance not
affected by the issue.